API/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
python-dotenv
requests
inflect
google-genai